refact/front/NextBtn.jsx: moves to src/components/buttons/NextBtn.jsx
